Refactor code to separeate Authentication settings into seprate files and move test constructor for IssuerSigningTokenProvider to test code
