+++ libs/host/Configuration/Options.cs
@@ -11,8 +11,8 @@ using System.Reflection;
 using System.Security.Cryptography.X509Certificates;
 using CommandLine;
 using Garnet.server;
-using Garnet.server.Auth;
 using Garnet.server.Auth.Aad;
+using Garnet.server.Auth.Settings;
 using Garnet.server.TLS;
 using Microsoft.Extensions.Logging;
 using Tsavorite.core;
