+++ libs/server/Auth/Aad/IssuerSigningTokenProvider.cs
@@ -31,7 +31,7 @@ namespace Garnet.server.Auth.Aad
 
         private readonly ILogger _logger;
 
-        private IssuerSigningTokenProvider(string authority, IReadOnlyCollection<SecurityKey> signingTokens, bool refreshTokens, ILogger logger)
+        protected IssuerSigningTokenProvider(string authority, IReadOnlyCollection<SecurityKey> signingTokens, bool refreshTokens = true, ILogger logger = null)
         {
             _authority = authority;
             if (refreshTokens)
@@ -108,17 +108,6 @@ namespace Garnet.server.Auth.Aad
 
             var signingTokens = RetrieveSigningTokens(authority);
             return new IssuerSigningTokenProvider(authority, signingTokens, refreshTokens: true, logger);
-        }
-
-        /// <summary>
-        /// [Used for testing]
-        /// Creates an instance of IssuerSigningTokenProvider without refresh of tokens and hence this doesn't need authority. 
-        /// 
-        /// </summary>
-        /// <param name="logger">The logger</param>
-        public static IssuerSigningTokenProvider Create(IReadOnlyCollection<SecurityKey> signingToken, ILogger logger)
-        {
-            return new IssuerSigningTokenProvider(string.Empty, signingToken, refreshTokens: false, logger);
         }
     }
 }
