+++ libs/server/Resp/ACLCommands.cs
@@ -7,6 +7,7 @@ using System.Text;
 using Garnet.common;
 using Garnet.server.ACL;
 using Garnet.server.Auth;
+using Garnet.server.Auth.Settings;
 using Microsoft.Extensions.Logging;
 
 namespace Garnet.server
