+++ test/Garnet.test.cluster/ClusterAadAuthTests.cs
@@ -5,15 +5,13 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
-using System.Text;
-using System.Threading.Tasks;
-using Garnet.server.Auth;
-using Garnet.server.Auth.Aad;
+using Garnet.server.Auth.Settings;
 using Microsoft.IdentityModel.Tokens;
 using NUnit.Framework;
 
 namespace Garnet.test.cluster
 {
+
     [TestFixture]
     [NonParallelizable]
     class ClusterAadAuthTests
@@ -21,6 +19,8 @@ namespace Garnet.test.cluster
         ClusterTestContext context;
 
         readonly HashSet<string> monitorTests = [];
+
+        private const string issuer = "https://sts.windows.net/975f013f-7f24-47e8-a7d3-abc4752bf346/";
 
         [SetUp]
         public void Setup()
@@ -41,7 +41,7 @@ namespace Garnet.test.cluster
         {
             var nodes = 2;
             var audience = Guid.NewGuid().ToString();
-            JwtTokenGenerator tokenGenerator = new JwtTokenGenerator(audience);
+            JwtTokenGenerator tokenGenerator = new JwtTokenGenerator(issuer, audience);
 
             var appId = Guid.NewGuid().ToString();
             var objId = Guid.NewGuid().ToString();
@@ -51,7 +51,7 @@ namespace Garnet.test.cluster
                 new Claim("appid", appId),
                 new Claim("http://schemas.microsoft.com/identity/claims/objectidentifier",objId),
             };
-            var authSettings = new AadAuthenticationSettings([appId], [audience], [audience], IssuerSigningTokenProvider.Create(new List<SecurityKey> { tokenGenerator.SecurityKey }, context.logger), true);
+            var authSettings = new AadAuthenticationSettings([appId], [audience], [issuer], new MockIssuerSigningTokenProvider(new List<SecurityKey> { tokenGenerator.SecurityKey }, context.logger), true);
 
             var token = tokenGenerator.CreateToken(tokenClaims, DateTime.Now.AddMinutes(10));
             ValidateConnectionsWithToken(objId, token, nodes, authSettings);
@@ -63,7 +63,7 @@ namespace Garnet.test.cluster
         {
             var nodes = 2;
             var audience = Guid.NewGuid().ToString();
-            JwtTokenGenerator tokenGenerator = new JwtTokenGenerator(audience);
+            JwtTokenGenerator tokenGenerator = new JwtTokenGenerator(issuer, audience);
 
             var appId = Guid.NewGuid().ToString();
             var objId = Guid.NewGuid().ToString();
@@ -75,7 +75,7 @@ namespace Garnet.test.cluster
                 new Claim("http://schemas.microsoft.com/identity/claims/objectidentifier", objId),
                 new Claim("groups", string.Join(',', groupIds)),
             };
-            var authSettings = new AadAuthenticationSettings([appId], [audience], [audience], IssuerSigningTokenProvider.Create(new List<SecurityKey> { tokenGenerator.SecurityKey }, context.logger), true);
+            var authSettings = new AadAuthenticationSettings([appId], [audience], [issuer], new MockIssuerSigningTokenProvider(new List<SecurityKey> { tokenGenerator.SecurityKey }, context.logger), true);
             var token = tokenGenerator.CreateToken(tokenClaims, DateTime.Now.AddMinutes(10));
             ValidateConnectionsWithToken(groupIds.First(), token, nodes, authSettings);
         }
