+++ test/Garnet.test.cluster/JwtTokenHelpers.cs
@@ -8,13 +8,24 @@ using System.Linq;
 using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
+using Garnet.server.Auth.Aad;
+using Microsoft.Extensions.Logging;
 using Microsoft.IdentityModel.Tokens;
 
 namespace Garnet.test.cluster
 {
+    internal class MockIssuerSigningTokenProvider : IssuerSigningTokenProvider
+    {
+        internal MockIssuerSigningTokenProvider(IReadOnlyCollection<SecurityKey> signingTokens, ILogger logger = null) : base(string.Empty, signingTokens, false, logger)
+        { }
+
+    }
+
     internal class JwtTokenGenerator
     {
         private readonly string Issuer;
+
+        private readonly string Audience;
 
         // Our random signing key - used to sign and validate the tokens
         public SecurityKey SecurityKey { get; }
@@ -25,16 +36,17 @@ namespace Garnet.test.cluster
         // the token handler we'll use to actually issue tokens
         public readonly JwtSecurityTokenHandler JwtSecurityTokenHandler = new();
 
-        internal JwtTokenGenerator(string issuer)
+        internal JwtTokenGenerator(string issuer, string audience)
         {
             Issuer = issuer;
+            Audience = audience;
             SecurityKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes("Test secret key for authentication and signing the token to be generated"));
             SigningCredentials = new SigningCredentials(SecurityKey, SecurityAlgorithms.HmacSha256);
         }
 
         internal string CreateToken(List<Claim> claims, DateTime expiryTime)
         {
-            return JwtSecurityTokenHandler.WriteToken(new JwtSecurityToken(Issuer, Issuer, claims, expires: expiryTime, signingCredentials: SigningCredentials));
+            return JwtSecurityTokenHandler.WriteToken(new JwtSecurityToken(Issuer, Audience, claims, expires: expiryTime, signingCredentials: SigningCredentials));
         }
 
     }
